SC-E-FS fuelCrossFeed=True and Fuel Gen properties
Changed the fuelCrossFeed=True to accommodate the changes to the RBDC and the Shuttle Body for cross feed from the tank made. *Enable Cross Feed on Tank in order to work*. Changed the FuelGen for the Shuttle to accept the Aerozine50/NTO fuel for EC generation. Original MP value was Ratio = 0.001875 and divided by two for Ratio = 0.0009375.